Separated card consts
